+++ scripts/install.sh
@@ -13,10 +13,11 @@ git clone https://github.com/WordPress/gutenberg.git
 cd gutenberg
 
 # checkout to version we want
-git checkout tags/v3.9.0
+git checkout tags/v4.2.0-rc.1
 
 # remove git references
-rm -rf .git
+# GUTENBERG HAS A SUBMODULE NOW
+# rm -rf .git
 
 # install gutenberg dependencies
 npm i