+++ src/js/gutenberg-overrides/edit-post/index.js
@@ -1,2 +1,0 @@
-export { PluginDocumentSidebarPanel } from './components/sidebar/document-sidebar';
-export * from 'gutenberg/edit-post?source=node_modules';