+++ src/js/gutenberg-overrides/packages/block-library/build-module/gallery/edit.js
@@ -33,16 +33,21 @@ import {
  */
 import GalleryImage from './gallery-image';
 
-import * as others from 'gutenberg/packages/block-library/build-module/gallery/edit?source=node_modules';
-
 const MAX_COLUMNS = 8;
 const linkOptions = [
   { value: 'attachment', label: __('Attachment Page') },
   { value: 'media', label: __('Media File') },
   { value: 'none', label: __('None') },
 ];
+const ALLOWED_MEDIA_TYPES = [ 'image' ];
 
-const { defaultColumnsNumber } = others;
+export function defaultColumnsNumber (attributes) {
+  return Math.min(3, attributes.images.length);
+}
+
+export const pickRelevantMediaFiles = image => {
+  return pick(image, [ 'alt', 'id', 'link', 'url', 'caption' ]);
+};
 
 class GalleryEdit extends Component {
   constructor (props) {
@@ -87,7 +92,7 @@ class GalleryEdit extends Component {
 
   onSelectImages (images) {
     this.props.setAttributes({
-      images: images.map(image => pick(image, [ 'alt', 'caption', 'id', 'link', 'url' ])),
+      images: images.map(image => pickRelevantMediaFiles(image)),
     });
   }
 
@@ -132,11 +137,12 @@ class GalleryEdit extends Component {
     const currentImages = this.props.attributes.images || [];
     const { noticeOperations, setAttributes } = this.props;
     mediaUpload({
-      allowedType: 'image',
+      allowedTypes: ALLOWED_MEDIA_TYPES,
       filesList: files,
       onFileChange: images => {
+        const imagesNormalized = images.map(image => pickRelevantMediaFiles(image));
         setAttributes({
-          images: currentImages.concat(images),
+          images: currentImages.concat(imagesNormalized),
         });
       },
       onError: noticeOperations.createErrorNotice,
@@ -169,7 +175,7 @@ class GalleryEdit extends Component {
           <Toolbar>
             <MediaUpload
               onSelect={ this.onSelectImages }
-              type="image"
+              allowedTypes={ ALLOWED_MEDIA_TYPES }
               multiple
               gallery
               value={ images.map(img => img.id) }
@@ -196,11 +202,11 @@ class GalleryEdit extends Component {
             className={ className }
             labels={ {
               title: __('Gallery'),
-              name: __('images'),
+              instructions: __('Drag images, upload new ones or select files from your library.'),
             } }
             onSelect={ this.onSelectImages }
             accept="image/*"
-            type="image"
+            allowedTypes={ ALLOWED_MEDIA_TYPES }
             multiple
             notices={ noticeUI }
             onError={ noticeOperations.createErrorNotice }
@@ -228,7 +234,7 @@ class GalleryEdit extends Component {
               help={ this.getImageCropHelp }
             />
             <SelectControl
-              label={ __('Link to') }
+              label={ __('Link To') }
               value={ linkTo }
               onChange={ this.setLinkTo }
               options={ linkOptions }
@@ -244,12 +250,12 @@ class GalleryEdit extends Component {
                 url={ img.url }
                 alt={ img.alt }
                 id={ img.id }
-                data={ img.data }
                 isSelected={ isSelected && this.state.selectedImage === index }
                 onRemove={ this.onRemoveImage(index) }
                 onSelect={ this.onSelectImage(index) }
                 setAttributes={ attrs => this.setImageAttributes(index, attrs) }
                 caption={ img.caption }
+                data={ img.data }
               />
             </li>
           )) }
@@ -274,4 +280,3 @@ class GalleryEdit extends Component {
 }
 
 export default withNotices(GalleryEdit);
-export * from 'gutenberg/packages/block-library/build-module/gallery/edit?source=node_modules';