+++ src/js/gutenberg-overrides/packages/block-library/build-module/image/edit.js
@@ -3,7 +3,6 @@
  */
 import React from 'react';
 import classnames from 'classnames';
-import ResizableBox from 're-resizable';
 import {
   get,
   isEmpty,
@@ -17,17 +16,19 @@ import {
  */
 import { __ } from '@wordpress/i18n';
 import { Component, Fragment } from '@wordpress/element';
-import { getBlobByURL, revokeBlobURL } from '@wordpress/blob';
+import { getBlobByURL, revokeBlobURL, isBlobURL } from '@wordpress/blob';
 import {
   Button,
   ButtonGroup,
   IconButton,
   PanelBody,
+  ResizableBox,
   SelectControl,
   TextControl,
   TextareaControl,
   Toolbar,
   withNotices,
+  ToggleControl,
 } from '@wordpress/components';
 import { withSelect } from '@wordpress/data';
 import {
@@ -55,24 +56,56 @@ const LINK_DESTINATION_NONE = 'none';
 const LINK_DESTINATION_MEDIA = 'media';
 const LINK_DESTINATION_ATTACHMENT = 'attachment';
 const LINK_DESTINATION_CUSTOM = 'custom';
+const ALLOWED_MEDIA_TYPES = [ 'image' ];
+
+export const pickRelevantMediaFiles = image => {
+  return pick(image, [ 'alt', 'id', 'link', 'url', 'caption' ]);
+};
+
+/**
+ * Is the URL a temporary blob URL? A blob URL is one that is used temporarily
+ * while the image is being uploaded and will not have an id yet allocated.
+ *
+ * @param {number=} id The id of the image.
+ * @param {string=} url The url of the image.
+ *
+ * @return {boolean} Is the URL a Blob URL
+ */
+const isTemporaryImage = (id, url) => ! id && isBlobURL(url);
+
+/**
+ * Is the url for the image hosted externally. An externally hosted image has no id
+ * and is not a blob url.
+ *
+ * @param {number=} id  The id of the image.
+ * @param {string=} url The url of the image.
+ *
+ * @return {boolean} Is the url an externally hosted url?
+ */
+const isExternalImage = (id, url) => url && ! id && ! isBlobURL(url);
 
 class ImageEdit extends Component {
   constructor (props) {
     super(props);
+    const { attributes } = props;
+
     this.updateAlt = this.updateAlt.bind(this);
     this.updateAlignment = this.updateAlignment.bind(this);
     this.onFocusCaption = this.onFocusCaption.bind(this);
     this.onImageClick = this.onImageClick.bind(this);
     this.onSelectImage = this.onSelectImage.bind(this);
+    this.onSelectURL = this.onSelectURL.bind(this);
     this.updateImageURL = this.updateImageURL.bind(this);
     this.updateWidth = this.updateWidth.bind(this);
     this.updateHeight = this.updateHeight.bind(this);
     this.updateDimensions = this.updateDimensions.bind(this);
     this.onSetCustomHref = this.onSetCustomHref.bind(this);
     this.onSetLinkDestination = this.onSetLinkDestination.bind(this);
+    this.toggleIsEditing = this.toggleIsEditing.bind(this);
 
     this.state = {
       captionFocused: false,
+      isEditing: ! attributes.url,
     };
   }
 
@@ -80,26 +113,26 @@ class ImageEdit extends Component {
     const { attributes, setAttributes } = this.props;
     const { id, url = '' } = attributes;
 
-    if (! id && url.indexOf('blob:') === 0) {
+    if (isTemporaryImage(id, url)) {
       const file = getBlobByURL(url);
 
       if (file) {
         mediaUpload({
           filesList: [ file ],
           onFileChange: ([ image ]) => {
-            setAttributes({ ...image });
+            setAttributes(pickRelevantMediaFiles(image));
           },
-          allowedType: 'image',
+          allowedTypes: ALLOWED_MEDIA_TYPES,
         });
       }
     }
   }
 
   componentDidUpdate (prevProps) {
-    const { id: prevID, url: prevUrl = '' } = prevProps.attributes;
+    const { id: prevID, url: prevURL = '' } = prevProps.attributes;
     const { id, url = '' } = this.props.attributes;
 
-    if (! prevID && prevUrl.indexOf('blob:') === 0 && id && url.indexOf('blob:') === -1) {
+    if (isTemporaryImage(prevID, prevURL) && ! isTemporaryImage(id, url)) {
       revokeBlobURL(url);
     }
 
@@ -120,8 +153,14 @@ class ImageEdit extends Component {
       });
       return;
     }
+
+    this.setState({
+      isEditing: false,
+    });
+
+    // GUTENBERG JS
     const toUpdate = {
-      ...pick(media, [ 'alt', 'id', 'caption', 'url' ]),
+      ...pickRelevantMediaFiles(media),
       width: undefined,
       height: undefined,
     };
@@ -157,6 +196,21 @@ class ImageEdit extends Component {
     this.props.setAttributes({
       linkDestination: value,
       href,
+    });
+  }
+
+  onSelectURL (newURL) {
+    const { url } = this.props.attributes;
+
+    if (newURL !== url) {
+      this.props.setAttributes({
+        url: newURL,
+        id: undefined,
+      });
+    }
+
+    this.setState({
+      isEditing: false,
     });
   }
 
@@ -222,9 +276,52 @@ class ImageEdit extends Component {
     ];
   }
 
+  toggleIsEditing () {
+    this.setState({
+      isEditing: ! this.state.isEditing,
+    });
+  }
+
   render () {
+    const { isEditing } = this.state;
     const { attributes, setAttributes, isLargeViewport, isSelected, className, maxWidth, noticeOperations, noticeUI, toggleSelection, isRTL } = this.props;
-    const { url, alt, caption, align, id, href, linkDestination, width, height, data } = attributes;
+    const { url, alt, caption, align, id, href, linkDestination, width, height, linkTarget, data } = attributes;
+    const isExternal = isExternalImage(id, url);
+
+    let toolbarEditButton;
+    if (url) {
+      if (isExternal) {
+        toolbarEditButton = (
+          <Toolbar>
+            <IconButton
+              className="components-icon-button components-toolbar__control"
+              label={ __('Edit image') }
+              onClick={ this.toggleIsEditing }
+              icon="edit"
+            />
+          </Toolbar>
+        );
+      }
+      else {
+        toolbarEditButton = (
+          <Toolbar>
+            <MediaUpload
+              onSelect={ this.onSelectImage }
+              allowedTypes={ ALLOWED_MEDIA_TYPES }
+              value={ id }
+              render={ ({ open }) => (
+                <IconButton
+                  className="components-toolbar__control"
+                  label={ __('Edit image') }
+                  icon="edit"
+                  onClick={ open }
+                />
+              ) }
+            />
+          </Toolbar>
+        );
+      }
+    }
 
     const controls = (
       <BlockControls>
@@ -232,48 +329,32 @@ class ImageEdit extends Component {
           value={ align }
           onChange={ this.updateAlignment }
         />
-
-        <Toolbar>
-          <MediaUpload
-            onSelect={ this.onSelectImage }
-            type="image"
-            value={ id }
-            render={ ({ open }) => (
-              <IconButton
-                className="components-toolbar__control"
-                label={ __('Edit image') }
-                icon="edit"
-                onClick={ open }
-              />
-            ) }
-          />
-        </Toolbar>
+        { toolbarEditButton }
       </BlockControls>
     );
 
-    if (! url) {
+    if (isEditing) {
+      const src = isExternal ? url : undefined;
       return (
         <Fragment>
           { controls }
           <MediaPlaceholder
             icon="format-image"
-            labels={ {
-              title: __('Image'),
-              name: __('an image'),
-            } }
             className={ className }
             onSelect={ this.onSelectImage }
+            onSelectURL={ this.onSelectURL }
             notices={ noticeUI }
             onError={ noticeOperations.createErrorNotice }
             accept="image/*"
-            type="image"
+            allowedTypes={ ALLOWED_MEDIA_TYPES }
+            value={ { id, src } }
           />
         </Fragment>
       );
     }
 
     const classes = classnames(className, {
-      'is-transient': 0 === url.indexOf('blob:'),
+      'is-transient': isBlobURL(url),
       'is-resized': !! width || !! height,
       'is-focused': isSelected,
     });
@@ -289,7 +370,7 @@ class ImageEdit extends Component {
             label={ __('Alt Text (Alternative Text)') }
             value={ alt }
             onChange={ this.updateAlt }
-            help={ __('Describe the purpose of the image. Leave empty if the image is not a key part of the content.') }
+            help={ __('Alternative text describes your image to people who can’t see it. Add a short description with its key details.') }
           />
           { ! isEmpty(availableSizes) && (
             <SelectControl
@@ -365,13 +446,21 @@ class ImageEdit extends Component {
             options={ this.getLinkDestinationOptions() }
             onChange={ this.onSetLinkDestination }
           />
-          <TextControl
-            label={ __('Link URL') }
-            value={ href || '' }
-            onChange={ this.onSetCustomHref }
-            placeholder={ ! isLinkURLInputDisabled ? 'https://' : undefined }
-            disabled={ isLinkURLInputDisabled }
-          />
+          { linkDestination !== LINK_DESTINATION_NONE && (
+            <Fragment>
+              <TextControl
+                label={ __('Link URL') }
+                value={ href || '' }
+                onChange={ this.onSetCustomHref }
+                placeholder={ ! isLinkURLInputDisabled ? 'https://' : undefined }
+                disabled={ isLinkURLInputDisabled }
+              />
+              <ToggleControl
+                label={ __('Open in New Tab') }
+                onChange={ () => setAttributes({ linkTarget: ! linkTarget ? '_blank' : undefined }) }
+                checked={ linkTarget === '_blank' } />
+            </Fragment>
+          ) }
         </PanelBody>
       </InspectorControls>
     );
@@ -462,11 +551,6 @@ class ImageEdit extends Component {
                     minHeight={ minHeight }
                     maxHeight={ maxWidth / ratio }
                     lockAspectRatio
-                    handleClasses={ {
-                      right: 'wp-block-image__resize-handler-right',
-                      bottom: 'wp-block-image__resize-handler-bottom',
-                      left: 'wp-block-image__resize-handler-left',
-                    } }
                     enable={ {
                       top: false,
                       right: showRightHandle,
@@ -494,7 +578,7 @@ class ImageEdit extends Component {
             <RichText
               tagName="figcaption"
               placeholder={ __('Write caption…') }
-              value={ caption || [] }
+              value={ caption }
               unstableOnFocus={ this.onFocusCaption }
               onChange={ value => setAttributes({ caption: value }) }
               isSelected={ this.state.captionFocused }