+++ src/js/gutenberg-overrides/packages/components/build-module/index.js
@@ -1,3 +1,0 @@
-
-export { default as CategorySelect } from './query-controls/category-select.js';
-export * from 'gutenberg/packages/components/build-module/index?source=node_modules';