+++ src/js/index.js
@@ -2,60 +2,139 @@
  * Internal dependencies
  */
 import './wp-init.js';
-import '../scss/style.scss';
+import './url-init.js';
+import './api-fetch-init.js';
 
 /**
  * WordPress dependencies
  */
-import * as blockLibrary from '@wordpress/block-library';
-import * as blocks from '@wordpress/blocks';
-import * as components from '@wordpress/components';
-import * as compose from '@wordpress/compose';
-import * as coreData from '@wordpress/core-data';
-import * as data from '@wordpress/data';
-import domReady from '@wordpress/dom-ready';
-import * as editor from '@wordpress/editor';
-import * as editPost from '@wordpress/edit-post';
+import * as autop from '@wordpress/autop';
+import * as blob from '@wordpress/blob';
+import * as blockSerializationDefaultParser from '@wordpress/block-serialization-default-parser';
+
+import * as escapeHtml from '@wordpress/escape-html';
 import * as element from '@wordpress/element';
+import * as isShallowEqual from '@wordpress/is-shallow-equal';
+import * as compose from '@wordpress/compose';
 import * as hooks from '@wordpress/hooks';
+import deprecated from '@wordpress/deprecated';
+import * as reduxRoutine from '@wordpress/redux-routine';
+
+import * as data from '@wordpress/data';
+import './packages/data.js';
+
+import * as dom from '@wordpress/dom';
 import * as i18n from '@wordpress/i18n';
+import * as shortcode from '@wordpress/shortcode';
+import * as blocks from '@wordpress/blocks';
+import domReady from '@wordpress/dom-ready';
+import * as a11y from '@wordpress/a11y';
+
+import * as url from '@wordpress/url';
+import apiFetch from '@wordpress/api-fetch';
+import './packages/api-fetch.js';
+
+import * as htmlEntities from '@wordpress/html-entities';
+import * as keycodes from '@wordpress/keycodes';
+import * as richText from '@wordpress/rich-text';
+import * as components from '@wordpress/components';
+
+import * as coreData from '@wordpress/core-data';
+
+import * as date from '@wordpress/date';
+import * as nux from '@wordpress/nux';
+import * as tokenList from '@wordpress/token-list';
+import * as viewport from '@wordpress/viewport';
+import * as wordcount from '@wordpress/wordcount';
+
+import { editor, oldEditor } from './packages/editor.js';
+
 import * as plugins from '@wordpress/plugins';
+import * as blockLibrary from '@wordpress/block-library';
+import * as editPost from '@wordpress/edit-post';
 
-// and styles
-import 'gutenberg/block-library/editor.scss';
-import 'gutenberg/block-library/theme.scss';
-import 'gutenberg/block-library/style.scss';
+// Style
+import 'gutenberg/packages/components/build-style/style.css';
+import 'gutenberg/packages/nux/build-style/style.css';
+import 'gutenberg/packages/editor/build-style/style.css';
+import 'gutenberg/packages/block-library/build-style/theme.css';
+import 'gutenberg/packages/block-library/build-style/editor.css';
+import 'gutenberg/packages/edit-post/build-style/style.css';
+import '../scss/style.scss';
 
-// Adding modules to wp global
-window.wp.blockLibrary = blockLibrary;
-window.wp.blocks = blocks;
-window.wp.components = components;
-window.wp.compose = compose;
-window.wp.coreData = coreData;
-window.wp.data = data;
-window.wp.domReady = domReady;
-window.wp.editor = {
-  ...window.wp.editor,
-  ...editor,
+// // Set global wp
+window.wp = {
+  apiFetch,
+  url,
+
+  autop,
+  blob,
+  blockSerializationDefaultParser,
+  escapeHtml,
+  element,
+  isShallowEqual,
+  compose,
+  hooks,
+  deprecated,
+  reduxRoutine,
+  data,
+  dom,
+  i18n,
+  shortcode,
+  blocks,
+  domReady,
+  a11y,
+  htmlEntities,
+  keycodes,
+  richText,
+  components,
+  coreData,
+  date,
+  nux,
+  tokenList,
+  viewport,
+  wordcount,
+  editor,
+  oldEditor,
+  plugins,
+  blockLibrary,
+  editPost,
 };
-window.wp.editPost = editPost;
-window.wp.element = element;
-window.wp.hooks = hooks;
-window.wp.i18n = i18n;
-window.wp.plugins = plugins;
 
 export {
-  blockLibrary,
-  blocks,
-  components,
-  compose,
-  coreData,
-  data,
-  domReady,
-  editor,
-  editPost,
+  apiFetch,
+  url,
+
+  autop,
+  blob,
+  blockSerializationDefaultParser,
+  escapeHtml,
   element,
+  isShallowEqual,
+  compose,
   hooks,
+  deprecated,
+  reduxRoutine,
+  data,
+  dom,
   i18n,
+  shortcode,
+  blocks,
+  domReady,
+  a11y,
+  htmlEntities,
+  keycodes,
+  richText,
+  components,
+  coreData,
+  date,
+  nux,
+  tokenList,
+  viewport,
+  wordcount,
+  editor,
+  oldEditor,
   plugins,
+  blockLibrary,
+  editPost,
 };