+++ src/js/packages/api-fetch.js
@@ -1,0 +1,5 @@
+import apiFetch from '@wordpress/api-fetch';
+
+// Middleware
+apiFetch.use(apiFetch.createNonceMiddleware(window.wpApiSettings.nonce));
+apiFetch.use(apiFetch.createRootURLMiddleware(window.wpApiSettings.root));