+++ src/js/gutenberg-overrides/packages/block-library/CHANGELOG.md
@@ -1,5 +1,11 @@
 # block-library changelog
 
+## 3.0.0
+
+### Changed
+
+- 'cover-image' to 'cover' ([packages/block-library/build-module/cover](https://github.com/front/gutenberg-js/blob/v3.0.0/src/js/gutenberg-overrides/packages/block-library/build-module/cover))
+
 ## 2.0.0 2018-08
 
 ### Changed